requirements.txt
flask
cryptography
pycryptodome
requests
gitpython